Seperated searchPlaces used in #31 for a better Cohesion and low Coupling
